Changed short to byte to save memory

+++ PR1-K17/src/main/java/com/k17/pr1k17/GameNumber.java
@@ -2,19 +2,19 @@ package com.k17.pr1k17;
 
 public class GameNumber
 {
-    private short value;
+    private byte value;
     private boolean isSelected;
 
-    public GameNumber(short value) {
+    public GameNumber(byte value) {
         this.value = value;
         this.isSelected = false;
     }
 
-    public short getValue() {
+    public byte getValue() {
         return value;
     }
 
-    public void setValue(short value) {
+    public void setValue(byte value) {
         this.value = value;
     }
